+++ backend/canisters/community/impl/src/updates/follow_thread.rs
@@ -17,7 +17,9 @@ fn follow_thread_impl(args: Args, state: &mut RuntimeState) -> OCResult {
     let user_id = member.user_id;
 
     if let Some(channel) = state.data.channels.get_mut(&args.channel_id) {
-        channel.chat.follow_thread(user_id, args.thread_root_message_index, now)?;
+        channel
+            .chat
+            .follow_thread(user_id, args.thread_root_message_index, now, &state.data.migrated_user_ids)?;
         state.mark_activity_for_user(user_id);
 
         if args.new_achievement && !member.user_type.is_bot() {