+++ backend/canisters/multi_user/impl/src/queries/deleted_message.rs
@@ -6,7 +6,9 @@ use user_canister::deleted_message::{Response::*, *};
 #[query(guard = "caller_is_hosted_user", msgpack = true)]
 fn deleted_message(args: Args) -> Response {
     match read_state(|state| {
-        state.with_caller_user(|my_index, user| user_core::queries::deleted_message(user, args, state.user_id(my_index)))
+        state.with_caller_user(|my_index, user| {
+            user_core::queries::deleted_message(user, args, state.user_id(my_index), &state.data.migrated_user_ids)
+        })
     }) {
         Ok(result) => Success(result),
         Err(error) => Error(error),